Fix memory corruption bugs in pgsql stored procedures.
These bug fixes are largely the result of adjusting buffer sizes and
some aspects of resource management, to make Valgrind stop complaining.
They're not particularly well thought out, and may create new problems.
For example, I was rather careless about freeing extra memory allocated
for temp buffers in some places, though it's not clear that it matters.

This work is partially based on changes from taiki-k's DBT-5 Github
repo, though it's still so rough in this revision that the final version
might end up having very little in common with that code.